+++ NEWS.md
@@ -1,3 +1,10 @@
+# multiprobit 0.2.4
+
+* Wrapper class `wp_model` for common methods interface for all
+  Wishart-like models
+* Basic posterior uncertainty approximations for the probit Sigma parameter
+  in the `summary()` method for `wm_estimate` objects.
+
 # multiprobit 0.2.3
 
 * First stable release